Add files via upload
Using HoloLens 2 to interact with virtual world and the real world simultaneously to learn signs and gestures for communication between people of all kinds non-verbally.

This technology can have many applications, from assisting individuals with disabilities to bridging the gap between people with language barriers. This tool will also be a good way to establish a universal form of communication across all types of individuals and cultures. 

This current build is extremely basic, incorporating the gestural functions of the HoloLens 2 in association with the current example object placed in the world, for demonstrative purposes. 

This version of the application needs the following:

- Azure AI features*
- Scene flow for the UI with text
- Graphical text
- Buttons activating gestural demonstration
- Sounds

Created build for a template for the true functions of the program. This will act as a demo as the rest of the application is built. The initial object used in this current build is a chair.

Current gestures include chair, human, and hello. 

True application would incorporate an extensive database of hand gestures and sign language gestures from all over the world, as it differs from country to country. This allows for more diverse learning and communication between various individuals non-verbally.

*Azure AI needs to be at least partially imported for recognition of objects. This may go into this build if time allows, otherwise it will likely be statically incorporated into the design of the application workflow. Since the device can notice objects in any case as it is allowed, this is really time based.